+++ tests/agenda_future_tasks_tests.py
@@ -12,3 +12,20 @@ def test_future_tasks_normal_flow():
     assert x.days[0].courses[0].items.materials == []
     assert x.days[0].courses[0].items.tasks[0].assignmentID == "549556"
     assert x.days[0].courses[0].items.tasks[0].description == "Toets 3. De koolstofcyclus in het systeem aarde pagina 42 - 47"
+
+    assert x.days[0].courses[0].items.tasks[0].endMomentTS is None
+
+
+def test_future_tasks_moved_task():
+    """
+    "Toets Heelal" was moved from 2023-11-20 -> 2023-11-27.
+    """
+    x: FutureTasks = FutureTasks()
+
+    assert x.days[0].date == date(2023, 11, 27)
+    assert x.days[0].courses[0].course_title == '1 - STEM1, Carla Verhoelst'
+    assert x.days[0].courses[0].items.materials == []
+    assert x.days[0].courses[0].items.tasks[0].assignmentID == "551612"
+    assert x.days[0].courses[0].items.tasks[0].description == 'Toets Heelal'
+
+    assert x.days[0].courses[0].items.tasks[0].endMomentTS == '318_20231127'